+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix comment update error when list item gets deleted
 * Markdown editor: Fix SVGs not rendered in math preview
 * Add design for HTB CWPE certification
+* Plugin `jira`: Export findings to Jira issues
 
 
 ## v2026.4 - 2026-01-23